feat: Refactor admin registration to unified user registration
