+++ server/src/identity_access_management_context/application/commands/__init__.py
@@ -3,7 +3,7 @@ from .update_user_command import UpdateUserCommand
 from .delete_user_command import DeleteUserCommand
 from .validate_user_token_command import ValidateUserTokenCommand
 from .admin_login_command import AdminLoginCommand
-from .register_admin_with_password_command import RegisterWithPasswordCommand
+from .register_with_password_command import RegisterWithPasswordCommand
 from .sso_login_command import SsoLoginCommand
 from .get_user_me_command import GetUserMeCommand
 from .refresh_access_token_command import RefreshAccessTokenCommand
