+++ server/src/identity_access_management_context/application/commands/register_with_password_command.py
